refactor: move _rclone_log_contains_not_ignored_errors to BackupService, simplify _rclone_copy_files_error_handler
This function is not utility function, as it is bound to BackupService logic, thus should be part of it, as private method. This also simplified _rclone_copy_files_error_handler.
